requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
python-dotenv
sentence-transformers
streamlit
langchain-core
fastapi
uvicorn
langchain-unstructured
unstructured